+++ src/main/java/tigerworkshop/webapphardwarebridge/dtos/LocalPrintDeviceActivationDTO.java
@@ -6,4 +6,5 @@ import lombok.Data;
 public class LocalPrintDeviceActivationDTO {
     private String shopId;
     private Long activationTaskId;
+    private String activationToken;
 }